requirements.txt
# AI助教系统 - 依赖文件
# 第10章：异常检测

# ============== 核心框架 ==============
streamlit>=1.28.0

# ============== 大模型API ==============
zhipuai>=2.0.0

# ============== 向量数据库 ==============
chromadb>=0.4.0

# ============== PDF处理 ==============
PyMuPDF>=1.23.0

# ============== 数据处理 ==============
numpy>=1.24.0
pandas>=2.0.0

# ============== 其他工具 ==============
python-dotenv>=1.0.0

